use mutex only in std
